rdtsc fixes and remove scatter/gather snappy in favor of a contiguous block and using libsnappy-c.. producer speed up to 1.3MM messages a sec at 200 byte message

+++ src/rd_rdtsc.h
@@ -36,9 +36,9 @@ unsigned long long rd_us_to_rdtsc(unsigned long long);
 
 /**
  * Returns number of microseconds in a tick interval.
- * Provides nanosecond granularity.
+ * Provides microsecond granularity;
  */
-double rd_rdtsc_scale(uint64_t);
+uint64_t rd_rdtsc_scale(uint64_t);
 
 /**
  * Returns number of microseconds in a tick interval
